Remove examples/tutorials. Examples have moved to their own repo.
PiperOrigin-RevId: 333794392
