+++ tensorflow/lite/experimental/examples/lstm/input_data.py
@@ -25,6 +25,7 @@ import collections
 import gzip
 import os
 
+
 import numpy
 from six.moves import urllib
 from six.moves import xrange  # pylint: disable=redefined-builtin
@@ -138,7 +139,7 @@ class _DataSet(object):
     Args:
       images: The images
       labels: The labels
-      fake_data: Ignore images and labels, use fake data.
+      fake_data: Ignore inages and labels, use fake data.
       one_hot: Bool, return the labels as one hot vectors (if True) or ints (if
         False).
       dtype: Output image dtype. One of [uint8, float32]. `uint8` output has
@@ -330,3 +331,4 @@ def read_data_sets(train_dir,
   test = _DataSet(test_images, test_labels, **options)
 
   return _Datasets(train=train, validation=validation, test=test)
+
